Move function to deal with pdfs to utils/
